+++ blocks/category-family/product-skeleton.js
@@ -1,0 +1,85 @@
+import {
+  div, p,
+} from '../../scripts/dom-builder.js';
+
+export const productSkeleton = div(
+  { class: 'coveo-skeleton flex flex-col lg:flex-row grid-rows-1 lg:grid-cols-5 gap-x-10 gap-y-4' },
+  div(
+    { class: 'col-span-1 border shadow rounded-lg w-full p-4 max-w-sm w-full' },
+    div(
+      { class: 'flex flex-col gap-y-4 animate-pulse' },
+      div({ class: 'w-2/4 h-7 bg-danaheratomicgrey-200 rounded [&:not(:first-child)]:opacity-40' }),
+      div({ class: 'w-3/4 h-4 bg-danaheratomicgrey-200 rounded [&:not(:first-child):even]:opacity-40' }),
+      div({ class: 'w-2/5 h-3 bg-danaheratomicgrey-200 rounded [&:not(:first-child):odd]:opacity-20' }),
+      div({ class: 'w-4/5 h-5 bg-danaheratomicgrey-200 rounded [&:not(:first-child):even]:opacity-40' }),
+    ),
+  ),
+  div(
+    { class: 'col-span-4 w-full' },
+    div({ class: 'max-w-xs bg-neutral-300 rounded-md p-4 animate-pulse mb-4' }),
+    div(
+      { class: 'grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-5' },
+      div(
+        { class: 'flex flex-col gap-y-2 animate-pulse' },
+        div({ class: 'h-72 rounded bg-danaheratomicgrey-200 opacity-500' }),
+        div({ class: 'w-2/4 h-7 bg-danaheratomicgrey-200 rounded [&:not(:first-child)]:opacity-40' }),
+        div(
+          { class: 'space-y-1' },
+          p({ class: 'w-3/4 h-4 bg-danaheratomicgrey-200 rounded [&:not(:first-child)]:opacity-40' }),
+          p({ class: 'w-2/5 h-3 bg-danaheratomicgrey-200 rounded [&:not(:first-child)]:opacity-20' }),
+          p({ class: 'w-4/5 h-5 bg-danaheratomicgrey-200 rounded [&:not(:first-child):even]:opacity-40' }),
+        ),
+        div(
+          { class: 'grid grid-cols-3 gap-4' },
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-2' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-2' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+        ),
+      ),
+      div(
+        { class: 'flex flex-col gap-y-2 animate-pulse' },
+        div({ class: 'h-72 rounded bg-danaheratomicgrey-200 opacity-500' }),
+        div({ class: 'w-2/4 h-7 bg-danaheratomicgrey-200 rounded [&:not(:first-child)]:opacity-40' }),
+        div(
+          { class: 'space-y-1' },
+          p({ class: 'w-3/4 h-4 bg-danaheratomicgrey-200 rounded [&:not(:first-child):even]:opacity-40' }),
+          p({ class: 'w-2/5 h-3 bg-danaheratomicgrey-200 rounded [&:not(:first-child):odd]:opacity-20' }),
+          p({ class: 'w-4/5 h-5 bg-danaheratomicgrey-200 rounded [&:not(:first-child):even]:opacity-40' }),
+        ),
+        div(
+          { class: 'grid grid-cols-3 gap-4' },
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-2' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-2' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+        ),
+      ),
+      div(
+        { class: 'flex flex-col gap-y-2 animate-pulse' },
+        div({ class: 'h-72 rounded bg-danaheratomicgrey-200 opacity-500' }),
+        div({ class: 'w-2/4 h-7 bg-danaheratomicgrey-200 rounded [&:not(:first-child)]:opacity-40' }),
+        div(
+          { class: 'space-y-1' },
+          p({ class: 'w-3/4 h-4 bg-danaheratomicgrey-200 rounded [&:not(:first-child):even]:opacity-40' }),
+          p({ class: 'w-2/5 h-3 bg-danaheratomicgrey-200 rounded [&:not(:first-child):odd]:opacity-20' }),
+          p({ class: 'w-4/5 h-5 bg-danaheratomicgrey-200 rounded [&:not(:first-child):even]:opacity-40' }),
+        ),
+        div(
+          { class: 'grid grid-cols-3 gap-4' },
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-2' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-2' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+          div({ class: 'h-2 bg-danaheratomicgrey-200 rounded col-span-1' }),
+        ),
+      ),
+    ),
+  ),
+);
